bds/requirements.txt
django==1.11.12
channels==1.1.8
djangorestframework-jwt
djangorestframework
unicodecsv
werkzeug
requests
django-cors-headers